backend/requirements.txt
# Backend dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.2.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
pydantic>=2.10.0

# Azure AI / OpenAI
openai>=1.82.0
azure-identity>=1.21.0

# Azure Cosmos DB
azure-cosmos>=4.9.0

# Web scraping / analysis
requests>=2.32.0
beautifulsoup4>=4.13.0

# Export services
markdown>=3.8
weasyprint>=63.1
python-docx>=1.1.0
